+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2022-08-08
+## [1.1.1] - 2022-08-08
+
+- Added TUI autolaunch config variable (default False).
+- Cleaned up HTML, and added Metadata show/hide button.
+
+## [1.1.0] - 2022-08-07
 
 - Refactored HTML code to fix several small but annoying issues.
 - Added new configuration console script `tuiconf` that allows user to change/store settings for: